refactor(options-typing): no longer a string, but complex object
