+++ src/components/Inputs/interface.ts
@@ -1,10 +1,12 @@
+import type { Option } from "../../services/select-options";
+
 export interface InputProps {
   id: string;
 }
 
 export interface SelectInputProps extends InputProps {
   label: string;
-  options: string[];
+  options: Option[];
 }
 
 export interface TextInputProps extends InputProps {
