+++ src/components/NewResponsibleModal/interface.ts
@@ -1,0 +1,10 @@
+export type ResponsibleData = {
+  name: string;
+  cpf: string;
+  birthDate: string;
+  civilState: string;
+  nis: string;
+  address: string;
+  email: string;
+  phone: string;
+}
