--- src/components/NewResponsibleModal/interface/interface.ts
@@ -1,11 +1,0 @@
-export type ResponsibleData = {
-    name: string;
-    cpf: string;
-    birthDate: string;
-    civilState: string;
-    nis: string;
-    address: string;
-    email: string;
-    phone: string;
-
-  }
